+++ objectbox/lib/objectbox.dart
@@ -8,6 +8,7 @@ export 'src/admin.dart' show Admin;
 export 'src/annotations.dart';
 export 'src/box.dart' show Box, PutMode;
 export 'src/common.dart';
+export 'src/modelinfo/enums.dart' show OBXSyncFlags;
 export 'src/native/query/vector_search_results.dart';
 export 'src/query.dart'
     show